Fix error in migration, replace bigint with integer.

+++ code/meandyoutube_app/db/migrate/20131027165823_create_users.rb
@@ -9,7 +9,7 @@ class CreateUsers < ActiveRecord::Migration
       t.string :first_name
       t.string :token
       t.string :refresh_token
-      t.bigint :expires_at
+      t.integer :expires_at
     end
   end
 end
